feat: Move ItemsSourceView implementation to base, support IReadOnlyList<T>
